requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.1
cycler==0.11.0
Django==4.0.3
django-extensions==3.1.5
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
fonttools==4.32.0
idna==3.3
kiwisolver==1.4.2
matplotlib==3.5.1
numpy==1.22.3
packaging==21.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pydotplus==2.0.2
pyit2fls==0.7.9
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
unicodecsv==0.14.1
urllib3==1.26.9
